search: move EisInferenceEndpointMetadata to inference-common
Moved the EisInferenceEndpointMetadata type to inference-common so it
can be shared by both inference-endpoints and inference plugins and stay
up-to-date.

Updated the connectors endpoint to use the type and attach metadata to
the connector when it's available. This also lead to fixing a bug where
the model creator was never set on the connector since it was using the
wrong value from the metadata.
