+++ x-pack/platform/packages/shared/ai-infra/inference-common/index.ts
@@ -157,6 +157,7 @@ export {
   defaultInferenceEndpoints,
   InferenceEndpointProvider,
   elasticModelIds,
+  type EisInferenceEndpointMetadata,
 } from './src/inference_endpoints';
 
 export {
