+++ x-pack/platform/plugins/shared/search_inference_endpoints/public/components/model_status/model_status_badge.tsx
@@ -6,7 +6,8 @@
  */
 
 import React from 'react';
-import { type EisInferenceEndpointMetadata, EisModelStatus } from '../../types';
+import type { EisInferenceEndpointMetadata } from '@kbn/inference-common';
+import { EisModelStatus } from '../../types';
 
 import { ModelPreviewBadge } from './model_preview_badge';
 import { ModelDeprecatedBadge } from './model_deprecated_badge';
